ai_core/requirements.txt
langchain
langchain-core
langchain_openai
langchain_community
python-dotenv
pydantic
numpy
httpx
asyncpg
cachetools
fastapi
uvicorn

starlette
PyJWT